Read xlsx file with multiple sheets
Convienent way to read xlsx into R. Require "readlx", a relatively fast tool to read xlsx files
takes 2 argument: (1) path of the file, (2) combine
if combine = T, it returns a single dataframe. It is often to see if the dataset is too big/ too many rows and required multiple spreadsheets. (! the data from different sheets must have to same structure in order to "rbind".
if combine = F (default), it just returns a list.